backend/requirements.txt
fastapi[all]>=0.118.0
langchain>=0.3.27
langchain-groq>=0.3.8
langchain-openai>=0.3.33
psycopg2-binary>=2.9.10
pydantic>=2.0.0
python-dotenv>=1.1.1
qlalchemy>=2.0.43
uvicorn>=0.37.0